requirements.txt.txt
streamlit 
google-generativeai 
langchain 
google-cloud-firestore 
python-dotenv
